+++ src/Prelude/Backprop.hs
@@ -33,6 +33,8 @@ module Prelude.Backprop (
   , maximum
   , traverse
   , toList
+  , mapAccumL
+  , mapAccumR
   -- * Functor and Applicative
   , fmap
   , (<$>)
@@ -42,24 +44,20 @@ module Prelude.Backprop (
   -- * Misc
   , fromIntegral
   , realToFrac
-  , coerce
+  , E.coerce
   ) where
 
 import           Numeric.Backprop
-import           Prelude             (Num(..), Fractional(..), Eq(..), Ord(..), Functor, Foldable, Traversable, Applicative, (.), ($))
-import qualified Control.Applicative as P
-import qualified Data.Coerce         as C
-import qualified Data.Foldable       as P
-import qualified Prelude             as P
+import           Prelude                   (Num(..), Fractional(..), Ord(..), Functor, Foldable, Traversable, Applicative)
+import qualified Numeric.Backprop.Explicit as E
+import qualified Prelude                   as P
+import qualified Prelude.Backprop.Explicit as E
 
 -- | Lifted 'P.sum'
 sum :: forall t a s. (Foldable t, Functor t, Backprop (t a), Backprop a, Num a, Reifies s W)
     => BVar s (t a)
     -> BVar s a
-sum = liftOp1 . op1 $ \xs ->
-    ( P.sum xs
-    , (P.<$ xs)
-    )
+sum = E.sum E.addFunc E.zeroFunc
 {-# INLINE sum #-}
 
 -- | Lifted 'P.pure'.
@@ -67,11 +65,7 @@ pure
     :: forall t a s. (Foldable t, Applicative t, Backprop (t a), Backprop a, Reifies s W)
     => BVar s a
     -> BVar s (t a)
-pure = liftOp1 . op1 $ \x ->
-    ( P.pure x
-    , P.foldl' add (zero x)
-    -- , P.foldl' add zero
-    )
+pure = E.pure E.addFunc E.zeroFunc E.zeroFunc
 {-# INLINE pure #-}
 
 -- | Lifted 'P.product'
@@ -79,11 +73,7 @@ product
     :: forall t a s. (Foldable t, Functor t, Backprop (t a), Backprop a, Fractional a, Reifies s W)
     => BVar s (t a)
     -> BVar s a
-product = liftOp1 . op1 $ \xs ->
-    let p = P.product xs
-    in ( p
-       , \d -> (\x -> p * d / x) P.<$> xs
-       )
+product = E.product E.addFunc E.zeroFunc
 {-# INLINE product #-}
 
 -- | Lifted 'P.length'.
@@ -91,10 +81,7 @@ length
     :: forall t a b s. (Foldable t, Backprop (t a), Backprop b, Num b, Reifies s W)
     => BVar s (t a)
     -> BVar s b
-length = liftOp1 . op1 $ \xs ->
-    ( P.fromIntegral (P.length xs)
-    , P.const (zero xs)
-    )
+length = E.length E.addFunc E.zeroFunc E.zeroFunc
 {-# INLINE length #-}
 
 -- | Lifted 'P.minimum'.  Undefined for situations where 'P.minimum' would
@@ -103,11 +90,7 @@ minimum
     :: forall t a s. (Foldable t, Functor t, Backprop a, Ord a, Backprop (t a), Reifies s W)
     => BVar s (t a)
     -> BVar s a
-minimum = liftOp1 . op1 $ \xs ->
-    let m = P.minimum xs
-    in  ( m
-        , \d -> (\x -> if x == m then d else zero x) P.<$> xs
-        )
+minimum = E.minimum E.addFunc E.zeroFunc
 {-# INLINE minimum #-}
 
 -- | Lifted 'P.maximum'.  Undefined for situations where 'P.maximum' would
@@ -116,11 +99,7 @@ maximum
     :: forall t a s. (Foldable t, Functor t, Backprop a, Ord a, Backprop (t a), Reifies s W)
     => BVar s (t a)
     -> BVar s a
-maximum = liftOp1 . op1 $ \xs ->
-    let m = P.maximum xs
-    in  ( m
-        , \d -> (\x -> if x == m then d else zero x) P.<$> xs
-        )
+maximum = E.maximum E.addFunc E.zeroFunc
 {-# INLINE maximum #-}
 
 -- | Lifted 'P.fmap'.  Lifts backpropagatable functions to be
@@ -130,7 +109,7 @@ fmap
     => (BVar s a -> BVar s b)
     -> BVar s (f a)
     -> BVar s (f b)
-fmap f = collectVar . P.fmap f . sequenceVar
+fmap = E.fmap E.addFunc E.addFunc E.zeroFunc E.zeroFunc E.zeroFunc
 {-# INLINE fmap #-}
 
 -- | Alias for 'fmap'.
@@ -149,10 +128,8 @@ traverse
     => (BVar s a -> f (BVar s b))
     -> BVar s (t a)
     -> BVar s (f (t b))
-traverse f = collectVar
-           . P.fmap collectVar
-           . P.traverse f
-           . sequenceVar
+traverse = E.traverse E.addFunc E.addFunc E.addFunc
+                      E.zeroFunc E.zeroFunc E.zeroFunc E.zeroFunc
 {-# INLINE traverse #-}
 
 -- | Lifted 'P.liftA2'.  Lifts backpropagatable functions to be
@@ -168,8 +145,8 @@ liftA2
     -> BVar s (f a)
     -> BVar s (f b)
     -> BVar s (f c)
-liftA2 f x y = collectVar $ f P.<$> sequenceVar x
-                              P.<*> sequenceVar y
+liftA2 = E.liftA2 E.addFunc E.addFunc E.addFunc
+                  E.zeroFunc E.zeroFunc E.zeroFunc E.zeroFunc
 {-# INLINE liftA2 #-}
 
 -- | Lifted 'P.liftA3'.  Lifts backpropagatable functions to be
@@ -186,18 +163,9 @@ liftA3
     -> BVar s (f b)
     -> BVar s (f c)
     -> BVar s (f d)
-liftA3 f x y z = collectVar $ f P.<$> sequenceVar x
-                                P.<*> sequenceVar y
-                                P.<*> sequenceVar z
+liftA3 = E.liftA3 E.addFunc E.addFunc E.addFunc E.addFunc
+                  E.zeroFunc E.zeroFunc E.zeroFunc E.zeroFunc E.zeroFunc
 {-# INLINE liftA3 #-}
-
--- | Coerce items inside a 'BVar'.
-coerce
-    :: forall a b s. C.Coercible a b
-    => BVar s a
-    -> BVar s b
-coerce = coerceVar
-{-# INLINE coerce #-}
 
 -- | Lifted conversion between two 'P.Integral' instances.
 --
@@ -206,8 +174,7 @@ fromIntegral
     :: (Backprop a, P.Integral a, Backprop b, P.Integral b, Reifies s W)
     => BVar s a
     -> BVar s b
-fromIntegral = liftOp1 . op1 $ \x ->
-    (P.fromIntegral x, P.fromIntegral)
+fromIntegral = E.fromIntegral E.addFunc E.zeroFunc
 {-# INLINE fromIntegral #-}
 
 -- | Lifted conversion between two 'Fractional' and 'P.Real' instances.
@@ -217,8 +184,7 @@ realToFrac
     :: (Backprop a, Fractional a, P.Real a, Backprop b, Fractional b, P.Real b, Reifies s W)
     => BVar s a
     -> BVar s b
-realToFrac = liftOp1 . op1 $ \x ->
-    (P.realToFrac x, P.realToFrac)
+realToFrac = E.realToFrac E.addFunc E.zeroFunc
 {-# INLINE realToFrac #-}
 
 -- | Lifted version of 'P.toList'.  Takes a 'BVar' of a 'Traversable' of
@@ -229,5 +195,30 @@ toList
     :: (Traversable t, Backprop a, Reifies s W)
     => BVar s (t a)
     -> [BVar s a]
-toList = toListOfVar P.traverse
+toList = E.toList E.addFunc E.zeroFunc
 {-# INLINE toList #-}
+
+-- | Lifted version of 'P.mapAccumL'.
+--
+-- @since 0.2.2.0
+mapAccumL
+    :: (Traversable t, Backprop b, Backprop c, Backprop (t c), Reifies s W)
+    => (BVar s a -> BVar s b -> (BVar s a, BVar s c))
+    -> BVar s a
+    -> BVar s (t b)
+    -> (BVar s a, BVar s (t c))
+mapAccumL = E.mapAccumL E.addFunc E.addFunc E.zeroFunc E.zeroFunc E.zeroFunc
+{-# INLINE mapAccumL #-}
+
+-- | Lifted version of 'P.mapAccumR'.
+--
+-- @since 0.2.2.0
+mapAccumR
+    :: (Traversable t, Backprop b, Backprop c, Backprop (t c), Reifies s W)
+    => (BVar s a -> BVar s b -> (BVar s a, BVar s c))
+    -> BVar s a
+    -> BVar s (t b)
+    -> (BVar s a, BVar s (t c))
+mapAccumR = E.mapAccumR E.addFunc E.addFunc E.zeroFunc E.zeroFunc E.zeroFunc
+{-# INLINE mapAccumR #-}
+